math: Fix UB in ldbl-128 llroundl
Building with --enable-ubasn triggers:

$ math/test-float128-llround
testing _Float128 (without inline functions)
UBSAN: Undefined behaviour in ../sysdeps/ieee754/float128/../ldbl-128/s_llroundl.c:65:37 left shift of 562949953421312 by 14 cannot be represented in type 'long long int'
Aborted

+++ sysdeps/ieee754/ldbl-128/s_llroundl.c
@@ -62,7 +62,7 @@ __llroundl (_Float128 x)
 	    result = (long long int) i0;
 	  else
 	    {
-	      result = ((long long int) i0 << (j0 - 48)) | (j >> (112 - j0));
+	      result = (i0 << (j0 - 48)) | (j >> (112 - j0));
 #ifdef FE_INVALID
 	      if (sign == 1 && result == LLONG_MIN)
 		/* Rounding brought the value out of range.  */
